backend/requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.1
sqlalchemy>=1.4.47
python-dotenv>=1.0.0
scikit-learn==1.3.2
pandas==2.1.3
numpy==1.26.2
langchain==0.0.350
openai==1.3.5
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
pydantic>=1.10.7
python-multipart>=0.0.6
boto3>=1.26.142
watchdog>=3.0.0
python-dateutil>=2.8.2
aiofiles>=23.1.0
psycopg2-binary>=2.9.6
jsonschema>=4.17.3
xmltodict>=0.13.0